change SuccessfulTransactionMessage to SuccessfulTransactionModal
