+++ src/components/v2/SuccessfulTransactionModal/index.tsx
@@ -1,0 +1,66 @@
+/** @jsxImportSource @emotion/react */
+import React from 'react';
+import BigNumber from 'bignumber.js';
+import Typography from '@mui/material/Typography';
+
+import { convertWeiToCoins } from 'utilities/common';
+import { TokenId } from 'types';
+import { BscLink } from '../BscLink';
+import { Icon, IconName } from '../Icon';
+import { IModalProps, Modal } from '../Modal';
+import { useStyles } from './styles';
+
+export interface ISuccessfulTransactionModalProps extends Omit<IModalProps, 'children'> {
+  title: string;
+  message: string;
+  transactionHash: string;
+  amount?: {
+    tokenSymbol: TokenId;
+    valueWei: BigNumber;
+  };
+  className?: string;
+}
+
+export const SuccessfulTransactionModal: React.FC<ISuccessfulTransactionModalProps> = ({
+  className,
+  title,
+  message,
+  amount,
+  transactionHash,
+  isOpened,
+  handleClose,
+}) => {
+  const styles = useStyles();
+
+  return (
+    <Modal isOpened={isOpened} handleClose={handleClose}>
+      <div className={className} css={styles.container}>
+        <Icon name="check" css={styles.headerIcon} />
+
+        <h3 css={styles.title}>{title}</h3>
+
+        <div css={styles.messageContainer}>
+          <Typography variant="small1" component="p">
+            {message}
+          </Typography>
+
+          {amount && (
+            <div css={styles.amountContainer}>
+              <Icon name={amount.tokenSymbol as IconName} css={styles.amountTokenIcon} />
+
+              <Typography variant="small1" component="span">
+                {convertWeiToCoins({
+                  value: amount.valueWei,
+                  tokenSymbol: amount.tokenSymbol,
+                  returnInReadableFormat: true,
+                })}
+              </Typography>
+            </div>
+          )}
+        </div>
+
+        <BscLink hash={transactionHash} />
+      </div>
+    </Modal>
+  );
+};
