+++ src/components/v2/SuccessfulTransactionModal/styles.ts
@@ -9,6 +9,7 @@ export const useStyles = () => {
       display: flex;
       flex-direction: column;
       align-items: center;
+      padding: ${theme.spacing(3, 0, 10)};
     `,
     headerIcon: css`
       width: ${theme.spacing(16)};
